[INLONG-8171][Manager] Extract public class
