+++ inlong-manager/manager-pojo/src/main/java/org/apache/inlong/manager/pojo/sort/node/LoadNodeProviderFactory.java
@@ -20,24 +20,24 @@ package org.apache.inlong.manager.pojo.sort.node;
 import org.apache.inlong.manager.common.enums.ErrorCodeEnum;
 import org.apache.inlong.manager.common.exceptions.BusinessException;
 import org.apache.inlong.manager.pojo.sort.node.base.LoadNodeProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.ClickHouseProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.DorisProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.ElasticsearchProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.GreenplumProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.HBaseProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.HDFSProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.HiveProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.HudiProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.IcebergProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.KafkaProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.KuduProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.MySQLProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.OracleProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.PostgreSQLProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.RedisProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.SQLServerProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.StarRocksProvider;
-import org.apache.inlong.manager.pojo.sort.node.load.TDSQLPostgreSQLProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.ClickHouseProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.DorisProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.ElasticsearchProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.GreenplumProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.HBaseProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.HDFSProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.HiveProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.HudiProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.IcebergProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.KafkaProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.KuduProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.MySQLProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.OracleProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.PostgreSQLProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.RedisProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.SQLServerProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.StarRocksProvider;
+import org.apache.inlong.manager.pojo.sort.node.provider.TDSQLPostgreSQLProvider;
 
 import java.util.ArrayList;
 import java.util.List;
