+++ inlong-manager/manager-pojo/src/main/java/org/apache/inlong/manager/pojo/sort/node/provider/HBaseProvider.java
@@ -15,53 +15,55 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.pojo.sort.node.load;
+package org.apache.inlong.manager.pojo.sort.node.provider;
 
 import org.apache.inlong.manager.common.consts.SinkType;
-import org.apache.inlong.manager.pojo.sink.SinkField;
-import org.apache.inlong.manager.pojo.sink.tdsqlpostgresql.TDSQLPostgreSQLSink;
+import org.apache.inlong.manager.pojo.sink.hbase.HBaseSink;
 import org.apache.inlong.manager.pojo.sort.node.base.LoadNodeProvider;
 import org.apache.inlong.manager.pojo.stream.StreamField;
 import org.apache.inlong.manager.pojo.stream.StreamNode;
 import org.apache.inlong.sort.protocol.FieldInfo;
 import org.apache.inlong.sort.protocol.node.LoadNode;
-import org.apache.inlong.sort.protocol.node.load.TDSQLPostgresLoadNode;
+import org.apache.inlong.sort.protocol.node.load.HbaseLoadNode;
 import org.apache.inlong.sort.protocol.transformation.FieldRelation;
+
+import com.google.common.collect.Lists;
 
 import java.util.List;
 import java.util.Map;
 
 /**
- * The Provider for creating TDSQLPostgreSQL load nodes.
+ * The Provider for creating HBase load nodes.
  */
-public class TDSQLPostgreSQLProvider implements LoadNodeProvider {
+public class HBaseProvider implements LoadNodeProvider {
 
     @Override
     public Boolean accept(String sinkType) {
-        return SinkType.TDSQLPOSTGRESQL.equals(sinkType);
+        return SinkType.HBASE.equals(sinkType);
     }
 
     @Override
-    public LoadNode createNode(StreamNode nodeInfo, Map<String, StreamField> constantFieldMap) {
-        TDSQLPostgreSQLSink tdsqlPostgreSQLSink = (TDSQLPostgreSQLSink) nodeInfo;
-        Map<String, String> properties = parseProperties(tdsqlPostgreSQLSink.getProperties());
-        List<SinkField> sinkFieldList = tdsqlPostgreSQLSink.getSinkFieldList();
-        List<FieldInfo> fieldInfos = parseFieldInfos(sinkFieldList, tdsqlPostgreSQLSink.getSinkName());
-        List<FieldRelation> fieldRelations = parseSinkFields(sinkFieldList, constantFieldMap);
-
-        return new TDSQLPostgresLoadNode(
-                tdsqlPostgreSQLSink.getSinkName(),
-                tdsqlPostgreSQLSink.getSinkName(),
+    public LoadNode createLoadNode(StreamNode nodeInfo, Map<String, StreamField> constantFieldMap) {
+        HBaseSink hbaseSink = (HBaseSink) nodeInfo;
+        Map<String, String> properties = parseProperties(hbaseSink.getProperties());
+        List<FieldInfo> fieldInfos = parseSinkFieldInfos(hbaseSink.getSinkFieldList(), hbaseSink.getSinkName());
+        List<FieldRelation> fieldRelations = parseSinkFields(hbaseSink.getSinkFieldList(), constantFieldMap);
+        return new HbaseLoadNode(
+                hbaseSink.getSinkName(),
+                hbaseSink.getSinkName(),
                 fieldInfos,
                 fieldRelations,
-                null,
+                Lists.newArrayList(),
                 null,
                 null,
                 properties,
-                tdsqlPostgreSQLSink.getJdbcUrl(),
-                tdsqlPostgreSQLSink.getUsername(),
-                tdsqlPostgreSQLSink.getPassword(),
-                tdsqlPostgreSQLSink.getSchemaName() + "." + tdsqlPostgreSQLSink.getTableName(),
-                tdsqlPostgreSQLSink.getPrimaryKey());
+                hbaseSink.getTableName(),
+                hbaseSink.getNamespace(),
+                hbaseSink.getZkQuorum(),
+                hbaseSink.getRowKey(),
+                hbaseSink.getBufferFlushMaxSize(),
+                hbaseSink.getZkNodeParent(),
+                hbaseSink.getBufferFlushMaxRows(),
+                hbaseSink.getBufferFlushInterval());
     }
 }
