+++ inlong-manager/manager-pojo/src/main/java/org/apache/inlong/manager/pojo/sort/node/provider/TubeMqProvider.java
@@ -15,55 +15,46 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.pojo.sort.node.extract;
+package org.apache.inlong.manager.pojo.sort.node.provider;
 
 import org.apache.inlong.manager.common.consts.SourceType;
 import org.apache.inlong.manager.pojo.sort.node.base.ExtractNodeProvider;
-import org.apache.inlong.manager.pojo.source.oracle.OracleSource;
+import org.apache.inlong.manager.pojo.source.tubemq.TubeMQSource;
 import org.apache.inlong.manager.pojo.stream.StreamNode;
 import org.apache.inlong.sort.protocol.FieldInfo;
-import org.apache.inlong.sort.protocol.constant.OracleConstant.ScanStartUpMode;
 import org.apache.inlong.sort.protocol.node.ExtractNode;
-import org.apache.inlong.sort.protocol.node.extract.OracleExtractNode;
-
-import org.apache.commons.lang3.StringUtils;
+import org.apache.inlong.sort.protocol.node.extract.TubeMQExtractNode;
 
 import java.util.List;
 import java.util.Map;
 
 /**
- * The Provider for creating Oracle extract nodes.
+ * The Provider for creating TubeMQ extract nodes.
  */
-public class OracleProvider implements ExtractNodeProvider {
+public class TubeMqProvider implements ExtractNodeProvider {
 
     @Override
     public Boolean accept(String sourceType) {
-        return SourceType.ORACLE.equals(sourceType);
+        return SourceType.TUBEMQ.equals(sourceType);
     }
 
     @Override
-    public ExtractNode createNode(StreamNode streamNodeInfo) {
-        OracleSource source = (OracleSource) streamNodeInfo;
-        List<FieldInfo> fieldInfos = parseFieldInfos(source.getFieldList(), source.getSourceName());
+    public ExtractNode createExtractNode(StreamNode streamNodeInfo) {
+        TubeMQSource source = (TubeMQSource) streamNodeInfo;
+        List<FieldInfo> fieldInfos = parseStreamFieldInfos(source.getFieldList(), source.getSourceName());
         Map<String, String> properties = parseProperties(source.getProperties());
 
-        ScanStartUpMode scanStartupMode = StringUtils.isBlank(source.getScanStartupMode())
-                ? null
-                : ScanStartUpMode.forName(source.getScanStartupMode());
-        return new OracleExtractNode(
+        return new TubeMQExtractNode(
                 source.getSourceName(),
                 source.getSourceName(),
                 fieldInfos,
                 null,
                 properties,
-                source.getPrimaryKey(),
-                source.getHostname(),
-                source.getUsername(),
-                source.getPassword(),
-                source.getDatabase(),
-                source.getSchemaName(),
-                source.getTableName(),
-                source.getPort(),
-                scanStartupMode);
+                source.getMasterRpc(),
+                source.getTopic(),
+                source.getSerializationType(),
+                source.getGroupId(),
+                source.getSessionKey(),
+                source.getTid());
     }
 }
